handle amounts from env

+++ src/controllers/GameController.ts
@@ -917,7 +917,7 @@ class GameController {
 									betStatus = "won"
 									winningAmount =
 										bet.betAmount *
-										parseInt((process.env.AMOUNT_JODI as string) ?? "90")
+										parseFloat((process.env.AMOUNT_JODI as string) ?? "90")
 								} else if (
 									bet.betNumber.startsWith("A") &&
 									bet.betNumber[1] === resultNumber[0]
@@ -926,7 +926,7 @@ class GameController {
 									betStatus = "won"
 									winningAmount =
 										bet.betAmount *
-										parseInt((process.env.AMOUNT_HARUP as string) ?? "9")
+										parseFloat((process.env.AMOUNT_HARUP as string) ?? "9")
 								} else if (
 									bet.betNumber.startsWith("B") &&
 									bet.betNumber[1] === resultNumber[1]
@@ -935,7 +935,7 @@ class GameController {
 									betStatus = "won"
 									winningAmount =
 										bet.betAmount *
-										parseInt((process.env.AMOUNT_HARUP as string) ?? "9")
+										parseFloat((process.env.AMOUNT_HARUP as string) ?? "9")
 								}
 
 								if (!walletUserId.includes(bet.userId)) {
